Corrected sample rate information.
git-svn-id: https://appletv.nanopi.net/svn/trunk@168 7e1782f7-7b3f-45d9-b215-e5100ed717cc

+++ SapphireMetaData.h
@@ -149,7 +149,7 @@ typedef enum {
 
 - (long long)size;
 - (float)duration;
-- (int)sampleRate;
+- (Float64)sampleRate;
 - (UInt32)audioFormatID;
 - (int)episodeNumber;
 - (int)seasonNumber;
